+++ Scoring/Interfaces/IScoringContext.cs
@@ -23,6 +23,6 @@ namespace Citolab.QTI.ScoringEngine.Interfaces
         List<XDocument> AssessmentmentResults { get; set; }
         ILogger Logger { get; set; }
 
-        //bool? ProcessParallel { get; set; }
+        bool? ProcessParallel { get; set; }
     }
 }