+++ release_notes.md
@@ -2,6 +2,7 @@
 * Refactored for better (batch) performence:
     * don't use reflection to find exectors
     * setup executor before processing. Which means no more xPath on test definitions per assessmentResult.
+* Added option to process results parallel
 
 ## 1.0.1
 * changed baseType: int to integer